# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "榕树平台API",
        "title": "cc fintech practices API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "rliu",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/dp": {
            "get": {
                "description": "查询用户所拥有的数字人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digital person"
                ],
                "summary": "查询数字人接口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分页查询页数，默认为1",
                        "name": "page_no",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页查询页大小，默认为10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "查询返回的排序字段，默认为创建时间",
                        "name": "order_field",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式，默认为倒序",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.GetDpResp"
                        }
                    }
                }
            }
        },
        "/dp/{dp_link}": {
            "delete": {
                "description": "删除用户所拥有的数字人信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digital person"
                ],
                "summary": "删除数字人接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "需删除的数字人id",
                        "name": "dp_link",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.CommResp"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "登录以获取token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登录接口",
                "parameters": [
                    {
                        "description": "用户账号与加密的用户密码",
                        "name": "user_account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.AuthReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.AuthResp"
                        }
                    }
                }
            }
        },
        "/pubkey": {
            "get": {
                "description": "获取rsa公钥",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authorization"
                ],
                "summary": "获取公钥接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.PubKeyResp"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "注册用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "注册用户接口",
                "parameters": [
                    {
                        "description": "用户名称、用户账户与加密的用户密码",
                        "name": "user_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.CommResp"
                        }
                    }
                }
            }
        },
        "/resource/{resource_link}": {
            "delete": {
                "description": "删除用户所拥有的素材库素材",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource lib"
                ],
                "summary": "删除素材库素材接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "素材连接，tone or image",
                        "name": "resource_link",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.CommResp"
                        }
                    }
                }
            }
        },
        "/resource/{resource_type}": {
            "get": {
                "description": "查询用户所拥有的 or 公共的素材库素材信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource lib"
                ],
                "summary": "查询素材库接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "素材类型，tone or image",
                        "name": "resource_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页查询页数，默认为1",
                        "name": "page_no",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页查询页大小，默认为10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否查询公共素材，默认为否",
                        "name": "is_public",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.GetResourceResp"
                        }
                    }
                }
            }
        },
        "/{file_type}/{file_name}": {
            "get": {
                "description": "下载文件的统一接口，数字人、封面图片、素材库素材均通过本接口下载",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "tags": [
                    "download"
                ],
                "summary": "下载文件接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "下载的文件类型",
                        "name": "file_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "下载的文件名称",
                        "name": "file_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schema.AuthReq": {
            "type": "object",
            "properties": {
                "decrypt_data": {
                    "type": "string"
                },
                "user_account": {
                    "type": "string"
                }
            }
        },
        "schema.AuthResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "schema.CommResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "schema.DpEntity": {
            "type": "object",
            "properties": {
                "cover_image_link": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "dp_link": {
                    "type": "string"
                },
                "dp_name": {
                    "type": "string"
                },
                "hot_score": {
                    "type": "integer"
                },
                "owner": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
        "schema.GetDpResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.DpEntity"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "schema.GetResourceResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ResourceEntity"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "schema.PubKeyResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "schema.RegisterReq": {
            "type": "object",
            "properties": {
                "decrypt_data": {
                    "type": "string"
                },
                "user_account": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "schema.ResourceEntity": {
            "type": "object",
            "properties": {
                "cover_image_link": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "resouce_id": {
                    "type": "string"
                },
                "resource_describe": {
                    "type": "string"
                },
                "resource_link": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        }
    }
}